requirements.txt
selenium
Pillow==9.1.0